requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.10
